Separate BrowserTest and add helper methods for inserting db objects
