+++ universal-application-tool-0.0.1/test/repository/WithPostgresContainer.java
@@ -6,6 +6,7 @@ import akka.stream.Materializer;
 import com.google.common.collect.ImmutableMap;
 import io.ebean.Ebean;
 import io.ebean.EbeanServer;
+import java.util.Optional;
 import models.Applicant;
 import models.Person;
 import models.Program;
@@ -16,6 +17,8 @@ import org.junit.BeforeClass;
 import play.Application;
 import play.db.ebean.EbeanConfig;
 import play.test.Helpers;
+import services.question.QuestionDefinition;
+import services.question.TextQuestionDefinition;
 
 public class WithPostgresContainer {
 
@@ -66,5 +69,31 @@ public class WithPostgresContainer {
     EbeanConfig config = app.injector().instanceOf(EbeanConfig.class);
     EbeanServer server = Ebean.getServer(config.defaultServer());
     server.truncate(Applicant.class, Person.class, Program.class, Question.class);
+  }
+
+  /**
+   * Convenience method for saving a program to the database.
+   *
+   * @param name the name of the program to store
+   * @return the {@link Program} model that was created
+   */
+  protected static Program insertProgram(String name) {
+    Program program = new Program(name, "description");
+    program.save();
+    return program;
+  }
+
+  /**
+   * Convenience method for saving a question to the database with the given ID.
+   *
+   * @param id the ID for the question to save
+   * @return the {@link Question} model that was created
+   */
+  protected Question insertQuestion(long id) {
+    QuestionDefinition definition =
+        new TextQuestionDefinition(id, 1L, "", "", "", ImmutableMap.of(), Optional.empty());
+    Question question = new Question(definition);
+    question.save();
+    return question;
   }
 }
